arm template for franka control stack
